+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.7.3 - Unreleased
 
+### Added
+* Add `autocomplete` arg for `argcomplete`.
+
 
 ## Version 0.7.2 - Released 2023-02-02
 ### Added